[refactor] 拆分 tools.ts
